requirements.txt
pydantic>=2.0.0
pydantic-ai>=0.1.0
pytest==7.3.1
black==23.3.0
flake8==6.0.0
python-dotenv==1.0.0